requirements.txt
streamlit>=1.28.0
streamlit-mic-recorder>=0.0.8
openai-whisper>=20231117
google-genai>=1.0.0
python-dotenv>=1.0.0
edge-tts>=6.1.0
torch>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
